NIFI-8013 Add env var to set remote repo in git persistence
This commit adds an environment variable NIFI_REGISTRY_GIT_REPO so that
the relevant configuration can be set when using the Docker image from
DockerHub.

The configuration is set only if the environment variable is set and not
empty since it's not a mandatory value to be set for the Registry to
operate.